修正:共通処理をconcernsに集約
