+++ app/controllers/homes_controller.rb
@@ -1,4 +1,5 @@
 class HomesController < ApplicationController
+  include Common
   skip_before_action :require_login
   before_action :set_q, only: %i[top]
   before_action :set_center_of_jp, only: %i[top]
