+++ app/controllers/mountains_controller.rb
@@ -1,4 +1,5 @@
 class MountainsController < ApplicationController
+  include Common
   skip_before_action :require_login
   before_action :set_q, only: %i[index]
   before_action :set_center_of_jp, only: %i[index]
